+++ cmd/calyptia/get_core_instance_test.go
@@ -77,7 +77,7 @@ func Test_newCmdGetAggregators(t *testing.T) {
 		cmd := newCmdGetAggregators(configWithMock(&ClientMock{
 			AggregatorsFunc: func(ctx context.Context, projectID string, params cloud.AggregatorsParams) (cloud.Aggregators, error) {
 				wantNoEq(t, nil, params.Last)
-				wantEq(t, uint64(2), *params.Last)
+				wantEq(t, uint(2), *params.Last)
 				return want, nil
 			},
 		}))
